requirements.txt
sarvamai
python-dotenv
sounddevice
numpy
requests
webrtcvad-wheels
noisereduce